upstream some util classes into araragi
